+++ CHANGELOG.md
@@ -73,6 +73,10 @@
 ### Removed
 - remove 40KiB download on demand option
 
+
+### added
+
+- voice messages
 <a id="1_40_4"></a>
 
 ## [1.40.4] - 2023-09-14